Feat: UI feedbacks (#3) - Users can edit/load their codes.

+++ app/controllers/rules_controller.rb
@@ -1,39 +1,66 @@
 class RulesController < ApplicationController
   def index
-    @rules = current_user.rules;
+    @rules = Rule.all;
+    session[:return_to] = rules_path()
   end
   def create
     @rule = Rule.new(rule_params)
     @rule.user = @current_user
+
     if @rule.save
+      flash[:success] = "Saved!"
       redirect_to edit_rule_path(@rule)
     else
-      raise 'Failed to save the rule.'
+      flash[:error] = 'Failed to save the rule.'
+      redirect_to making_rule_path()
     end
+
+
   end
   def edit
     @rule = Rule.find(params[:id])
+
+    if @rule.user != current_user
+      flash[:error] = "You cannot edit others' rule."
+      redirect_to session.delete(:return_to)
+    end
+
   end
   def update
     @rule = Rule.find(params[:id])
 
-    if @rule.update(rule_params)
-      redirect_to making_rule_path(@rule)
+    if @rule.user == current_user
+      if @rule.update(rule_params)
+        flash[:success] = "Saved!"
+        redirect_to session.delete(:return_to)
+      else
+        flash[:error] = 'Failed to update the rule.'
+        redirect_to session.delete(:return_to)
+      end
     else
-      raise 'Failed to update the rule.'
+      flash[:error] = "You cannot edit others' rule."
+      redirect_to session.delete(:return_to)
     end
 
   end
   def destroy
 
     @rule = Rule.find(params[:id])
-    if @rule.destroy
-      redirect_to making_rule_path()
+    if @rule.user == current_user
+      if @rule.destroy
+        flash[:success] = "Destroyed!"
+        redirect_to rules_path()
+      else
+        flash[:error] = "Failed to destroy the rule."
+        redirect_to rules_path()
+      end
+
     else
-      raise 'Failed to destroy the rule.'
+      flash[:error] = "You cannot destroy others' rule."
+      redirect_to rules_path()
     end
 
-    redirect_to making_rule_path()
+
   end
 
 private
